Make certified mixed refinement a peer production option
Constraint: CMRC must coexist with Method-C, Red-Green, and HARP-DV without relabeling peers.
Rejected: Treat existing refinement backends as legacy | conflicts with the product contract.
Confidence: high
Scope-risk: broad
Directive: Preserve the final [40°,80°] interval proof and atomic artifact publication gates.
Tested: cargo test --workspace; cargo check --workspace; cargo clippy --workspace --all-targets -- -D warnings; GUI tests; mixed CMRC acceptance; real IGBP NXP80 acceptance.

+++ rust/earthmesh_cli/src/area_judge_types/source.rs
@@ -45,7 +45,7 @@ pub struct AreaJudgeDomainInitializationReport {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct AreaJudgeSeaOrLandReport {
     pub seaorland: Vec<Vec<bool>>,
-    pub sum_land_grid: i32,
+    pub sum_land_grid: u64,
 }
 
 /// Binary landtype class used by `MOD_Area_judge.F90` when building `seaorland`.
